Remove old benchmark test
